is_staff and select user to be staff in my website (Done)

+++ ShowRoom/cars/views.py
@@ -41,6 +41,11 @@ def all_cars(request):
         'color_filter': color_filter,
     })
 def create_car(request):
+    # Check if the user is staff
+    if not request.user.is_staff:
+        messages.error(request, "Only staff can add a car.", "alert-warning")
+        return redirect("main:index_view")
+    
     if request.method == 'POST':
         form = CarForm(request.POST, request.FILES)
         if form.is_valid():
@@ -48,7 +53,7 @@ def create_car(request):
             messages.success(request, "Car has been successfully added!")
             return redirect('cars:all_cars')
         else:
-            messages.error(request, "Error adding car. Please check the form.")
+            messages.error(request, "Error adding car. Please check the form and try again.")
     else:
         form = CarForm()
     return render(request, 'cars/create_car.html', {'form': form})
@@ -60,6 +65,9 @@ def car_detail(request, car_id):
 
 def update_car(request, car_id):
     car = get_object_or_404(Car, id=car_id)
+    if not request.user.is_staff:
+        messages.error(request, "Only staff can update a car.", "alert-warning")
+        return redirect('main:index_view')
     if request.method == 'POST':
         form = CarForm(request.POST, request.FILES, instance=car)
         if form.is_valid():
@@ -67,7 +75,7 @@ def update_car(request, car_id):
             messages.success(request, "Car has been successfully updated!")
             return redirect('cars:car_detail', car_id=car.id)
         else:
-            messages.error(request, "Error updating car. Please check the form.")
+            messages.error(request, f"Error updating car: {form.errors}. Please check the form.")
     else:
         form = CarForm(instance=car)
     return render(request, 'cars/update_car.html', {'form': form, 'car': car})
@@ -75,6 +83,10 @@ def update_car(request, car_id):
 
 def delete_car(request, car_id):
     car = get_object_or_404(Car, id=car_id)
+    # Permission check
+    if not request.user.is_staff:
+        messages.warning(request, "Only staff can delete cars.", "alert-warning")
+        return redirect("main:home_view")
     if request.method == 'POST':
         car_name = car.name
         car.delete()
